requirements.txt
torch==1.10.0
numpy==1.20.3
gym==0.18.3
gym-super-mario-bros==7.3.0
nes-py==8.1.8
opencv-python-headless==4.5.3.56
matplotlib==3.4.2
Pillow==8.2.0
